Fix circular dependency (#301)
between Signaling.js and PeerConnection.js
