+++ packages/millicast-sdk/docs-translations/en/utils/Codecs.json
@@ -1,0 +1,5 @@
+{
+    "<p>Enum of Millicast supported Video codecs</p>": "<p>Enum of Millicast supported Video codecs</p>",
+    "<p>Only available in Safari</p>": "<p>Only available in Safari</p>",
+    "<p>Enum of Millicast supported Audio codecs</p>": "<p>Enum of Millicast supported Audio codecs</p>"
+}
