+++ packages/millicast-sdk/src/Publish.js
@@ -3,7 +3,8 @@ import reemit from 're-emitter'
 import { atob } from 'Base64'
 import Logger from './Logger'
 import BaseWebRTC from './utils/BaseWebRTC'
-import Signaling, { signalingEvents, VideoCodec } from './Signaling'
+import Signaling, { signalingEvents } from './Signaling'
+import { VideoCodec } from './utils/Codecs'
 import PeerConnection, { webRTCEvents } from './PeerConnection'
 import FetchError from './utils/FetchError'
