+++ packages/millicast-sdk/src/Signaling.js
@@ -2,6 +2,7 @@ import EventEmitter from 'events'
 import TransactionManager from 'transaction-manager'
 import Logger from './Logger'
 import SdpParser from './utils/SdpParser'
+import { VideoCodec } from './utils/Codecs'
 import PeerConnection from './PeerConnection'
 
 const logger = Logger.get('Signaling')
@@ -11,36 +12,6 @@ export const signalingEvents = {
   connectionError: 'wsConnectionError',
   connectionClose: 'wsConnectionClose',
   broadcastEvent: 'broadcastEvent'
-}
-
-/**
- * Enum of Millicast supported Video codecs
- * @readonly
- * @enum {String}
- * @property {String} VP8
- * @property {String} VP9
- * @property {String} H264
- * @property {String} AV1
- * @property {String} H265 - Only available in Safari
- */
-export const VideoCodec = {
-  VP8: 'vp8',
-  VP9: 'vp9',
-  H264: 'h264',
-  AV1: 'av1',
-  H265: 'h265'
-}
-
-/**
- * Enum of Millicast supported Audio codecs
- * @readonly
- * @enum {String}
- * @property {String} OPUS
- * @property {String} MULTIOPUS
- */
-export const AudioCodec = {
-  OPUS: 'opus',
-  MULTIOPUS: 'multiopus'
 }
 
 /**
