+++ packages/millicast-sdk/src/utils/Codecs.js
@@ -1,0 +1,29 @@
+/**
+ * Enum of Millicast supported Video codecs
+ * @readonly
+ * @enum {String}
+ * @property {String} VP8
+ * @property {String} VP9
+ * @property {String} H264
+ * @property {String} AV1
+ * @property {String} H265 - Only available in Safari
+ */
+export const VideoCodec = {
+  VP8: 'vp8',
+  VP9: 'vp9',
+  H264: 'h264',
+  AV1: 'av1',
+  H265: 'h265'
+}
+
+/**
+ * Enum of Millicast supported Audio codecs
+ * @readonly
+ * @enum {String}
+ * @property {String} OPUS
+ * @property {String} MULTIOPUS
+ */
+export const AudioCodec = {
+  OPUS: 'opus',
+  MULTIOPUS: 'multiopus'
+}
