+++ packages/millicast-sdk/tests/features/step-definitions/Publish.steps.js
@@ -1,7 +1,8 @@
 import { loadFeature, defineFeature } from 'jest-cucumber'
 import Publish from '../../../src/Publish'
 import PeerConnection from '../../../src/PeerConnection'
-import Signaling, { VideoCodec } from '../../../src/Signaling'
+import Signaling from '../../../src/Signaling'
+import { VideoCodec } from '../../../src/utils/Codecs'
 import './__mocks__/MockRTCPeerConnection'
 import './__mocks__/MockMediaStream'
 import './__mocks__/MockBrowser'
